requirements.txt
Django==5.2.4
Pillow==11.3.0